Refactor: add Table parent class
